requirements.txt
google_api_python_client==2.141.0
google_auth_oauthlib==1.2.1
gsheets==0.6.1
huggingface_hub==0.24.6
imageio==2.35.0
matplotlib==3.7.5
numpy==1.24.4
openai_clip==1.0.1
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==10.4.0
protobuf==5.27.3
pycocoevalcap==1.2
pytorch_lightning==2.3.3
scipy==1.10.1
spacy==3.7.5
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.5
openai-clip==1.0.1
seaborn
openpyxl